+++ dsmrreader/__init__.py
@@ -1,5 +1,5 @@
 from django.utils.version import get_version
 
-VERSION = (4, 4, 0, 'beta', 4)
+VERSION = (4, 4, 0, 'beta', 5)
 
 __version__ = get_version(VERSION)